.Initial changes from 22.03.2020 00:22                 commity po requestach
